resume_parser/requirements.txt
flask==2.0.1
pymupdf==1.18.14
spacy==3.2.0
python-docx==0.8.11
textract==1.6.5
python-pptx==0.6.21